(wm) add support for detaching rows
This allows for adding logical partitions by splitting the row chain
into subgroups that can be dragged around individually. Actual drag-
group follows selection for the time being, though euclidian distance
from mouse_xy to closest row.bg should probably be considered.

This needs the relink_image fix in upstream arcan to not look terrible,
and there are probably more edge conditions to be squished when it comes
to autopanning.